fixed modal buttons behaviour, both buttons are hidden by default and with no default text. Changed the background color of the secondaryButton. Changed the position of the Modal Component in the project thtree > Now in the folder components/ui/Modal
